add otp file
