+++ tools/auth.js
@@ -11,9 +11,7 @@ const {
 	MISSING_HEADER,
 	DEACTIVATED_USER,
 	INVALID_CAPTCHA,
-	INVALID_OTP_CODE,
 	INVALID_PASSWORD,
-	TOKEN_OTP_MUST_BE_ENABLED,
 	TOKEN_NOT_FOUND,
 	TOKEN_REVOKED,
 	MULTIPLE_API_KEY,
@@ -42,9 +40,8 @@ const {
 	HMAC_TOKEN_KEY
 } = require(`${SERVER_PATH}/constants`);
 const rp = require('request-promise');
-const { getKitSecrets, getKitConfig, getFrozenUsers, getNetworkKeySecret } = require('./common');
+const { getKitSecrets, getFrozenUsers, getNetworkKeySecret } = require('./common');
 const dbQuery = require('./database/query');
-const otp = require('otp');
 const bcrypt = require('bcryptjs');
 const { getModel } = require('./database/model');
 const { client } = require('./database/redis');
@@ -54,6 +51,7 @@ const { sendEmail } = require(`${SERVER_PATH}/mail`);
 const { MAILTYPE } = require(`${SERVER_PATH}/mail/strings`);
 const { loggerAuth } = require(`${SERVER_PATH}/config/logger`);
 const moment = require('moment');
+const { checkUserOtpActive } = require('./otp');
 
 //Here we setup the security checks for the endpoints
 //that need it (in our case, only /protected). This
@@ -427,69 +425,6 @@ const checkCaptcha = (captcha = '', remoteip = '') => {
 		});
 };
 
-const generateOtp = (secret, epoch = 0) => {
-	const options = {
-		name: getKitConfig().api_name,
-		secret,
-		epoch
-	};
-	const totp = otp(options).totp();
-	return totp;
-};
-
-const verifyOtp = (userSecret, userDigits) => {
-	const serverDigits = [generateOtp(userSecret, 30), generateOtp(userSecret), generateOtp(userSecret, -30)];
-	return serverDigits.includes(userDigits);
-};
-
-const hasUserOtpEnabled = (id) => {
-	return dbQuery.findOne('user', {
-		where: { id },
-		attributes: ['otp_enabled']
-	}).then((user) => {
-		return user.otp_enabled;
-	});
-};
-
-const verifyUserOtpCode = (user_id, otp_code) => {
-	return dbQuery.findOne('otp code', {
-		where: {
-			used: true,
-			user_id
-		},
-		attributes: ['id', 'secret'],
-		order: [['updated_at', 'DESC']]
-	})
-		.then((otpCode) => {
-			return verifyOtp(otpCode.secret, otp_code);
-		})
-		.then((validOtp) => {
-			if (!validOtp) {
-				throw new Error(INVALID_OTP_CODE);
-			}
-			return true;
-		});
-};
-
-const verifyOtpBeforeAction = (user_id, otp_code) => {
-	return hasUserOtpEnabled(user_id).then((otp_enabled) => {
-		if (otp_enabled) {
-			return verifyUserOtpCode(user_id, otp_code);
-		} else {
-			return true;
-		}
-	});
-};
-
-const checkOtp = (userId) => {
-	return hasUserOtpEnabled(userId).then((otp_enabled) => {
-		if (otp_enabled) {
-			throw new Error('OTP is already enabled');
-		}
-		return findUserOtp(userId);
-	});
-};
-
 const checkAdminIp = (whiteListedIps = [], ip = '') => {
 	if (whiteListedIps.length === 0) {
 		return true; // no ip restriction for admin
@@ -549,89 +484,8 @@ const issueToken = (
 	return token;
 };
 
-/*
-	Function generate the otp secret.
-	Return the otp secret.
- */
-const generateOtpSecret = () => {
-	const seed = otp({
-		name: getKitConfig().api_name
-	});
-	return seed.secret;
-};
-
-/*
-	Function to find the user otp code, should take one parameter:
-	Param 1(integer): user id
-	Return a promise with the otp code row from the db.
- */
-const findUserOtp = (user_id) => {
-	return dbQuery.findOne('otp code', {
-		where: {
-			used: false,
-			user_id
-		},
-		attributes: ['id', 'secret']
-	});
-};
-
-/*
-	Function to create a user otp code, should take one parameter:
-	Param 1(integer): user id
-	Return a promise with the otp secret created.
- */
-const createOtp = (user_id) => {
-	const secret = generateOtpSecret();
-	return getModel('otp code').create({
-		user_id,
-		secret
-	})
-		.then((otpCode) => otpCode.secret);
-};
-
 const validatePassword = (userPassword, inputPassword) => {
 	return bcrypt.compare(inputPassword, userPassword);
-};
-
-/*
-  Function to find update the uset otp_enabled field,
-  should take two parameter:
-
-  Param 1(integer): user id
-  Param 2(boolean): otp_enabled
-
-  Return a promise with the updated user.
- */
-const updateUserOtpEnabled = (id, otp_enabled = false, transaction) => {
-	return dbQuery.findOne('user', {
-		where: { id },
-		attributes: ['id', 'otp_enabled']
-	}).then((user) => {
-		return user.update(
-			{ otp_enabled },
-			{ fields: ['otp_enabled'], transaction }
-		);
-	});
-};
-
-/*
-	Function to set used to true in the user otp code and update the user and set otp_enabled to true, should take one parameter:
-	Param 1(integer): user id
-	Return a promise with the user updated.
- */
-const setActiveUserOtp = (user_id) => {
-	return getModel('sequelize').transaction((transaction) => {
-		return findUserOtp(user_id)
-			.then((otp) => {
-				return otp.update(
-					{ used: true },
-					{ fields: ['used'], transaction }
-				);
-			})
-			.then(() => {
-				return updateUserOtpEnabled(user_id, true, transaction);
-			});
-	});
 };
 
 const getResetPasswordCode = (code) => {
@@ -720,27 +574,6 @@ const changeUserPassword = (email, oldPassword, newPassword) => {
 		.then((user) => {
 			return user.update({ password: newPassword });
 		});
-};
-
-const userHasOtpEnabled = (userId) => {
-	return require('./user').getUserByKitId(userId)
-		.then((user) => {
-			return user.otp_enabled;
-		});
-};
-
-const checkUserOtpActive = (userId, otpCode) => {
-	return all([
-		require('./user').getUserByKitId(userId),
-		verifyOtpBeforeAction(userId, otpCode)
-	]).then(([user, validOtp]) => {
-		if (!user.otp_enabled) {
-			throw new Error(TOKEN_OTP_MUST_BE_ENABLED);
-		} else if (!validOtp) {
-			throw new Error(INVALID_OTP_CODE);
-		}
-		return;
-	});
 };
 
 const createHmacSignature = (secret, verb, path, expires, data = '') => {
@@ -918,26 +751,15 @@ module.exports = {
 	verifyNetworkHmacToken,
 	userScopeIsValid,
 	isValidPassword,
-	hasUserOtpEnabled,
 	userIsDeactivated,
 	checkCaptcha,
-	verifyOtpBeforeAction,
-	verifyOtp,
-	checkOtp,
 	findToken,
 	issueToken,
 	validatePassword,
-	generateOtp,
-	generateOtpSecret,
-	findUserOtp,
-	setActiveUserOtp,
-	updateUserOtpEnabled,
-	createOtp,
 	sendResetPasswordCode,
 	resetUserPassword,
 	changeUserPassword,
 	getUserKitHmacTokens,
-	userHasOtpEnabled,
 	createUserKitHmacToken,
 	deleteUserKitHmacToken,
 	checkHmacSignature,
